+++ Plugin/Interfaces/IPluginInitialLoad.cs
@@ -1,0 +1,7 @@
+namespace UniTASPlugin.Interfaces;
+
+public interface IPluginInitialLoad
+{
+    void OnInitialLoad();
+    bool FinishedOperation { get; }
+}
